# TodoList/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  title: TodoList API
  version: 1.0.0
  description: A simple TodoList API
paths:
  /todos:
    get:
      summary: Get all todos
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Todo'
    post:
      summary: Create a new todo
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: todo
          in: body
          required: true
          schema:
            $ref: '#/components/schemas/Todo'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/components/schemas/Todo'
  /todos/{id}:
    get:
      summary: Get a todo by ID
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/components/schemas/Todo'
        '404':
          description: Not Found
    put:
      summary: Update a todo
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - name: todo
          in: body
          required: true
          schema:
            $ref: '#/components/schemas/Todo'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/components/schemas/Todo'
        '404':
          description: Not Found
    delete:
      summary: Delete a todo
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
components:
  schemas:
    Todo:
      type: object
      required:
        - title
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        completed:
          type: boolean
          default: false